A few suggestions
Perhaps also rename to "w3c-cg" as it's already in the namespace "openactive"?

"minutes will be posted to the group's public mailing list, or to a GitHub issue if the group uses GitHub." <- are we making a decision between these? Shall we pick one, or leave it to the community to decide? (I've seen schema.org make good use of GitHub Issues for purposes other than "Issues".